Bin: rewrite update-gitea script, so it download correct version now

+++ configs/bin/updaters/update-gitea.sh
@@ -1,41 +1,73 @@
 #!/usr/bin/env bash
 
+# ~ Config section begin ~
+
+DIR="/home/git/gitea"
+FILE="gitea"
+
+# ~ Config section end ~
+
 fail() {
 	echo "Update failed"
 	exit 1
 }
 
-echo "Update starting!"
+echo "Starting..."
 
-DOWNLOAD_SCRIPT="$(dirname "$(readlink -f "$0")")/downloaders/download-latest-release-from-github.sh"
+RELEASES_FILE=$(mktemp)
 
-DIR="/home/git/gitea"
-FILE="gitea"
-REGEX="browser_download_url.*gitea-.*-linux-amd64\"$"
-OPTS="--dir=$DIR --out=$FILE.new"
+curl --silent --output "$RELEASES_FILE" https://api.github.com/repos/go-gitea/gitea/releases
 
-if [ -f "$DOWNLOAD_SCRIPT" ]; then
-	# shellcheck disable=SC1090
-	if ! . "$DOWNLOAD_SCRIPT" "go-gitea/gitea" "$REGEX" "$OPTS"; then
-		fail
-	fi
+LAST_REPO_VERSION=$(grep -oP '(?<="tag_name": "v)(.*)(?=")' "$RELEASES_FILE" | sort -Vr | head -n1)
+LAST_VERSION_URL=$(grep -oP "(?<=\s\"browser_download_url\": \")https:\/\/github\.com\/go-gitea\/gitea\/releases\/download\/v$LAST_REPO_VERSION\/gitea-$LAST_REPO_VERSION-linux-amd64(?=\")" "$RELEASES_FILE")
+
+rm "$RELEASES_FILE"
+
+URL_REGEX="(https?|ftp|file)://[-A-Za-z0-9\+&@#/%?=~_|!:,.;]*[-A-Za-z0-9\+&@#/%=~_|]"
+if ! [[ "$LAST_VERSION_URL" =~ $URL_REGEX ]]; then
+	echo "Error: wrong regex: grep result is not url:"
+	echo "$LAST_VERSION_URL"
+
+	exit 1
+fi
+
+LOCAL_GITEA_VERSION=$("$DIR"/"$FILE" -v | grep -oP "(?<=Gitea version )(.*)(?= built)")
+COMPARED_VERSION=$(printf "%s\n%s" "$LOCAL_GITEA_VERSION" "$LAST_REPO_VERSION" | sort -Vr | head -n1)
+
+echo "Local gitea version: $LOCAL_GITEA_VERSION"
+echo "Latest gitea version: $LAST_REPO_VERSION"
+
+if [ "$LOCAL_GITEA_VERSION" != "$COMPARED_VERSION" ]; then
+	echo "Starting update!"
 else
-	echo "Error: Download script not found"
-	fail
+	echo "Update unneded!"
+fi
+
+echo "Downloading..."
+if ! aria2c --console-log-level=error --dir="$DIR" --out="$FILE.new" "$LAST_VERSION_URL"; then
+	echo "Downloading failed"
+	exit 1
 fi
 
 printf "Stopping Gitea serevice..."
 if ! systemctl stop gitea.service; then
 	fail
 fi
-printf " ok\n"
+printf " ✔\n"
+
+printf "Backing up old Gitea bin..."
+rm "$DIR/$FILE.bak" 2>/dev/null
+if ! cp "$DIR/$FILE" "$DIR/$FILE.bak"; then
+	fail
+fi
+printf " ✔\n"
 
 printf "Moving Gitea bin..."
 rm "$DIR/$FILE"
 if ! mv "$DIR/$FILE.new" "$DIR/$FILE"; then
 	fail
 fi
-printf " ok\n"
+printf " ✔\n"
 
 printf "Fixing permissions..."
 if ! chmod +x "$DIR/$FILE"; then
@@ -44,12 +76,18 @@ fi
 if ! chown -R git:git "$DIR/$FILE"; then
 	fail
 fi
-printf " ok\n"
+if ! chmod +x "$DIR/$FILE.bak"; then
+	fail
+fi
+if ! chown -R git:git "$DIR/$FILE.bak"; then
+	fail
+fi
+printf " ✔\n"
 
 printf "Starting Gitea serevice..."
 if ! systemctl start gitea.service; then
 	fail
 fi
-printf " ok\n"
+printf " ✔\n"
 
 echo "Update completed!"
